JS: Delete old query and update qhelp
